build: generate C headers like other *-sys crates
Generate C headers in the same manner as CMake / Autotools, but inside `build.rs`. The generated headers are written to a directory inside `$OUT_DIR` provided by Cargo.

This has no effect when producing a standalone binary for use in a C/C++ application, but when MesaLink is used in a Rust application, it allows other `-sys` crates to access the MesaLink headers during compilation for any other C code that may need them.

This is being added for the sake of `curl-sys`, but is a general-purpose solution.